pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omnigraph-agent"
version = "0.1.0"
description = "OmniGraph Agent for knowledge graph context building and NL→SPARQL generation"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "sparqlwrapper>=1.8.5",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rdflib>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
omnigraph-context-builder = "omnigraph_agent.context_builder.cli:main"

[project.entry-points."mcp.servers"]
omnigraph = "omnigraph_agent.mcp:create_server"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

